+++ Spoke.Unity/SpokeUnityBootstrap.cs
@@ -10,12 +10,13 @@ namespace Spoke {
 /// Static bootstrap for Spoke in Unity.
 /// Initializes SpokeError logging to use Unity's Debug.LogError.
 /// That's where Spoke logs internal errors by default.
-/// Also configures the default Spoke logger to use the UnitySpokeLogger.
+/// Also configures the default Spoke logger: UnityContext doubles as the logger,
+/// and one with no host behaviour logs to the console without pinging an object.
 /// </summary>
 #if UNITY_EDITOR
     [InitializeOnLoad]
 #endif
-    public static class SpokeUnityBootstrap {
+    internal static class SpokeUnityBootstrap {
         static bool isInitialized;
 
         static SpokeUnityBootstrap() 
@@ -26,7 +27,7 @@ namespace Spoke {
             if (isInitialized) return;
             isInitialized = true;
             SpokeError.Log = (msg, ex) => Debug.LogError($"[Spoke] {msg}\n{ex}");
-            SpokeError.DefaultLogger = new UnitySpokeLogger();
+            SpokeError.DefaultLogger = new UnityContext();
         }
     }
 }
